+++ lib/Property/ICalendar/DateTime.php
@@ -368,15 +368,17 @@ class DateTime extends Property {
 
         $messages = parent::validate($options);
         $valueType = $this->getValueType();
-        $value = $this->getValue();
+        $values = $this->getParts();
         try {
-            switch($valueType) {
-                case 'DATE' :
-                    DateTimeParser::parseDate($value);
-                    break;
-                case 'DATE-TIME' :
-                    DateTimeParser::parseDateTime($value);
-                    break;
+            foreach($values as $value) {
+                switch($valueType) {
+                    case 'DATE' :
+                        DateTimeParser::parseDate($value);
+                        break;
+                    case 'DATE-TIME' :
+                        DateTimeParser::parseDateTime($value);
+                        break;
+                }
             }
         } catch (\LogicException $e) {
             $messages[] = [