+++ tests/VObject/Recur/EventIterator/ByMonthInDailyTest.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Sabre\VObject;
+namespace Sabre\VObject\Recur;
 
 use
+    Sabre\VObject\Reader,
     DateTime;
 
-class RecurrenceIteratorByMonthInDailyTest extends \PHPUnit_Framework_TestCase {
+class ByMonthInDailyTest extends \PHPUnit_Framework_TestCase {
 
     /**
      * This tests the expansion of dates with DAILY frequency in RRULE with BYMONTH restrictions